two numbers problem in cpp

+++ Introduction to C++ for DSA/CodeForce_Problem/Sheet_1/D_Difference.cpp
@@ -1,11 +1,10 @@
-#include<bits/stdc++.h>
+#include <bits/stdc++.h>
 using namespace std;
 
-int main()
-{
-    int a,b,c,d;
-    cin>>a>>b>>c>>d;
-    int x = (a*b)-(c*d);
-    cout<<"Difference = "<<x<<endl; 
+int main() {
+    int a, b, c, d;
+    cin >> a >> b >> c >> d;
+    long long int x = (long long int)a * b - (long long int)c * d;
+    cout << "Difference = " << x << endl;
     return 0;
 }
